added notes in the main2 js file
4-29-24 Managed to translate the main game codes paradigm over to OOP style and implemented a title screen
which preceeds the game. Added a button in the title screen to enable the user to go to the game when they press the button.
Downloaded more asset images from the Google Drive folder made by Rania, she also resized the images to fit
the dimensions of the game properly, thank you bby girl.  Deleted the old main JS file because its not necessary anymore.