new screen and extra code
6-6-24
added a click anywhere screen that preceeds the title screen, the idea is that once the user clicks on this black screen with text, then music will start playing on the title screen immediately.
before this, the user had to click somewhere on the title screen to trigger the background music which is not normal.
Added an event that detects when user presses pointer on the game background and causes title screen to start