made audio and scene changes
6-11-24
Successfully managed to implement the title screen music which plays throughout option and credits screens too.  When the user enters the game scene, the title music ends,
but if the user decides to go back to the title screen (Effectively resetting the azetude game) then the title music is played once again from the beginning.  The user
can still hear the title music playing as they explore the options and credit scenes as before.  Started working on the options screen MUTE button and its functionality.